Adding support for affix random generation (-a)
there is a new flag -a which enables or
disables affix random string. The idea
being that if you generate a complete
random data - it takes a while. So with
-a, it generates data once and then just
mutate it one byte at a time. This is
enough for parquet file to get a sensible
size - without marking the data same as
before.