FIX: affix (-a) and TPC-DS generator error
TPC-DA data generator generators random sized strings
between [0, -s]. With affix only one string is generated
and at a random place, a char is substituted. This leads
to errors where there is a size mis-match between the
fixed affixed string and the size of the new string,
which is passed.

For now, we fix it by ignoring -a for TPC-DS data
generation, and printing and warning.