BUGFIX: Array[Byte] reference in data generation
The way Array[Byte] was generated, it was returning
reference (da!) to the dataset. That means when we
mutate values of the Array[Byte] all rows were
modified. Hence, we write it out the parquet
discards the column and just stores it once.

the fix is : always retrun a new array object. The
same is done for the StringBuilder where mkString
returns a new object.